// apps/backend/src/generated/schemas/bookmark-summary.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Pantry API
 * OpenAPI spec version: 0.0.0
 */
import type { BookmarkId } from "./bookmark-id";
import type { BookmarkNote } from "./bookmark-note";
import type { BookmarkTitle } from "./bookmark-title";
import type { BookmarkUrl } from "./bookmark-url";

export interface BookmarkSummary {
  id: BookmarkId;
  url: BookmarkUrl;
  title: BookmarkTitle;
  note?: BookmarkNote;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}