// apps/backend/src/generated/schemas/create-bookmark-request.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Pantry API
 * OpenAPI spec version: 0.0.0
 */
import type { BookmarkNote } from "./bookmark-note";
import type { BookmarkTitle } from "./bookmark-title";
import type { BookmarkUrl } from "./bookmark-url";
import type { TagInput } from "./tag-input";

export interface CreateBookmarkRequest {
  url: BookmarkUrl;
  title?: BookmarkTitle;
  note?: BookmarkNote;
  /** @maxItems 20 */
  tags?: TagInput[];
}